a little more cleanup

+++ src/org/usfirst/frc2619/MathUtil.java
@@ -4,7 +4,7 @@ public class MathUtil {
 	
 	public static double delinearize(double input, int power)
 	{		
-		return Math.pow(input, (double)power);
+		return Math.pow(input, power);
 	}
 	
 	public static double deadbandCheck(double input, double db){
